+++ admin/src/Field/CustomadminviewsField.php
@@ -42,6 +42,7 @@ class CustomadminviewsField extends ListField
 	 */
 	protected function getOptions()
 	{
+		// Get the databse object.
 		$db = Factory::getDBO();
 		$query = $db->getQuery(true);
 		$query->select($db->quoteName(array('a.id','a.system_name'),array('id','customadminview_system_name')));
@@ -50,10 +51,13 @@ class CustomadminviewsField extends ListField
 		$query->order('a.system_name ASC');
 		$db->setQuery((string)$query);
 		$items = $db->loadObjectList();
-		$options = array();
+		$options = [];
 		if ($items)
 		{
-			$options[] = Html::_('select.option', '', 'Select an option');
+			if ($this->multiple === false)
+			{
+				$options[] = Html::_('select.option', '', Text::_('COM_COMPONENTBUILDER_SELECT_AN_OPTION'));
+			}
 			foreach($items as $item)
 			{
 				$options[] = Html::_('select.option', $item->id, $item->customadminview_system_name);