+++ admin/src/View/Compiler/HtmlView.php
@@ -18,6 +18,7 @@ use Joomla\CMS\Layout\FileLayout;
 use Joomla\CMS\MVC\View\HtmlView as BaseHtmlView;
 use Joomla\CMS\Plugin\PluginHelper;
 use Joomla\CMS\Toolbar\ToolbarHelper;
+use Joomla\CMS\User\User;
 use Joomla\CMS\Document\Document;
 use VDM\Component\Componentbuilder\Administrator\Helper\HeaderCheck;
 use VDM\Component\Componentbuilder\Administrator\Helper\ComponentbuilderHelper;
@@ -27,7 +28,6 @@ use Joomla\CMS\Layout\LayoutHelper;
 use VDM\Joomla\Utilities\ArrayHelper;
 use VDM\Joomla\Utilities\FormHelper;
 use VDM\Joomla\Utilities\StringHelper;
-use Joomla\CMS\User\User;
 
 // No direct access to this file
 \defined('_JEXEC') or die;